Corrected the commentRegEx
According to PEP8 (https://www.python.org/dev/peps/pep-0008/#inline-comments): "An inline comment is a comment on the same line as a statement. Inline comments should be separated by at least two spaces from the statement. They should start with a # and a single space.". Opy was looking only for "#" to detect inline comments which generated problems as noted in this issue https://github.com/QQuick/Opy/issues/10 as well as others. 
This proposed fix would assume the python code was written following the most basic style guide available.